backend/requirements.txt
fastapi==0.78.0
gunicorn==20.1.0
pydantic==1.9.1
starlette==0.19.1
uvicorn==0.17.6
PyYAML==6.0
